LayoutTests/plugins/private-browsing-mode-2-expected.txt

Tests that NPNVprivateModeBool is supported by the WebKit plugin view. This test is for WebKit platforms that wish to support NPNVprivateModeBool but do not wish to implement the preference change listener required to support a cachedPrivateBrowsingEnabled property similar to the one provided by Safari and tested for in private-browsing-mode.html

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS plugin1.privateBrowsingEnabled is false
PASS plugin1.privateBrowsingEnabled is true
PASS plugin2.privateBrowsingEnabled is true
PASS plugin1.privateBrowsingEnabled is false
PASS plugin2.privateBrowsingEnabled is false
